✨ Implement navigation bar
The changes add new components and dependencies to the project. The
NavBar component is added to the Layout component to provide a
navigation bar to the application. The Logo component is created to
display the name of the author. The NavLink component is created to
provide a way to navigate between pages. The routes and social media
constants are added to provide a centralized way to manage the paths and
links used in the application. Finally, styles are added too.

487