✨ Implement Home page
The changes include adding a profile picture, brand message, paragraph,
and resume section to the Home page. The styles for the Home, Layout,
and AnimatedText components were also added. The color variable was
added to the globals.scss file, and the body color was set to dark. The
layout class was added to the root element in the Layout component.

683