✨ Add Profile component to about page
The Profile component is added to the about page to display information
about the author. The shadow.scss file now has mixins for box-shadow
elevation, which can be used to add depth to elements. The z-index.scss
file now has a z-index scale and applications, which can be used to
manage the stacking order of elements.

1047